requirements.txt
gradio
transformers
torch
sentencepiece
soundfile
openai-whisper
gTTS